pyproject.toml
[project]
name = "semshift"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "nltk>=3.9.2",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.2.2",
    "tqdm>=4.67.3",
    "transformers>=5.1.0",
]
